Report blocksize of 4096 for all zvol devices, similar to Advanced Format disks
Reporting a blocksize of 512 is more compatible (e.g. when cloning to/from physical devices that report a blocksize of 512). However taking advantage of the 'advanced format' 4k blocksize will reduce IO's and increase performance.
When used this way, the OS is presented with a usable device, and the zvol 'volblocksize' property serves the same purpose as 'recordsize' for datasets.

Would be nice to have an additional creation-time property to control the logical blocksize per zvol.